camera-stream-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pillow==11.1.0